@@ -1,30 +1,26 @@
 {
-  "name": "web",
-  "homepage": "web/html",
-  "version": "1.0.0",
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@types/styled-components": "^5.1.18",
-    "framer-motion": "^5.5.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.3.1",
-    "react-laag": "^2.0.3",
-    "styled-components": "^5.3.3"
-  },
-  "devDependencies": {
-    "@types/node": "^17.0.2",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@vitejs/plugin-react": "^1.0.7",
-    "autoprefixer": "^10.4.0",
-    "postcss": "^8.4.5",
-    "tailwindcss": "^3.0.7",
-    "typescript": "^4.4.4",
-    "vite": "^2.7.2"
-  }
+	"name": "radialmenu",
+	"homepage": "html",
+	"private": true,
+	"version": "0.1.0",
+	"scripts": {
+		"start": "vite",
+		"start:game": "vite build --watch --emptyOutDir",
+		"build": "tsc && vite build --emptyOutDir",
+		"preview": "vite preview"
+	},
+	"dependencies": {
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0"
+	},
+	"devDependencies": {
+		"@types/react": "^18.0.25",
+		"@types/react-dom": "^18.0.9",
+		"@vitejs/plugin-react": "^2.2.0",
+		"autoprefixer": "^10.4.15",
+		"postcss": "^8.4.29",
+		"tailwindcss": "^3.3.3",
+		"typescript": "^4.9.3",
+		"vite": "^3.2.4"
+	}
 }
