@@ -1,4 +1,4 @@
-import {isEnvBrowser} from "./misc";
+import { isEnvBrowser } from './misc';
 
 interface DebugEvent<T = any> {
   action: string;
@@ -13,16 +13,16 @@ interface DebugEvent<T = any> {
  * @param timer - How long until it should trigger (ms)
  */
 export const debugData = <P>(events: DebugEvent<P>[], timer = 1000): void => {
-  if (process.env.NODE_ENV === "development" && isEnvBrowser()) {
+  if (import.meta.env.MODE === 'development' && isEnvBrowser()) {
     for (const event of events) {
       setTimeout(() => {
         window.dispatchEvent(
-          new MessageEvent("message", {
+          new MessageEvent('message', {
             data: {
               action: event.action,
               data: event.data,
             },
-          })
+          }),
         );
       }, timer);
     }
