@@ -1,7 +1,15 @@
+/** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: ['./src/**/*.{js,jsx,ts,tsx}'],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
+	content: ['./index.html', './src/**/*.{js,ts,jsx,tsx}'],
+	theme: {
+		extend: {
+			fontFamily: {
+				NotoSans: ['Noto Sans JP', 'sans-serif'],
+			},
+			colors: {
+				primary: 'rgba(23, 23, 23, 0.70)',
+			},
+		},
+	},
+	plugins: [],
 };
