@@ -3,10 +3,10 @@ import react from '@vitejs/plugin-react';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()],
-  base: './',
-  build: {
-    outDir: './html',
-  },
-  publicDir: 'public',
+	plugins: [react()],
+	base: './',
+	build: {
+		outDir: '../html',
+	},
+	publicDir: 'public',
 });
